model/requirements.txt
Flask==2.0.1
tensorflow==2.4.1
Pillow==8.2.0
google-cloud-storage==1.38.0
numpy==1.19.5